OpenHarmony port: nopCommerce 5.00 working tree rebased on upstream develop
Working tree snapshot from the OHOS sandbox, rebased onto upstream
develop (b8a280ff9, 2026-09-07) so the branch carries real upstream
history. Diff vs parent shows only local changes:

- int32->int64 ID migration across Core/Data/Services/Web/Plugins/Tests
- SQLite + Oracle data providers; Yitter/TinyId id generators + tests
- net11 SDK pinning (global.json) and net10/net11 build adaptations
- OHOS sandbox tooling, docs, AGENTS.md, run scripts
- re-adds Nop.Plugin.Payments.Manual (removed upstream)
- .gitignore: exclude local agent state (.omo/, .arts/)

Line endings/BOM follow upstream; upstream-only additions since the
snapshot were retained; local build artifacts (sqlite dbs, deps.json)
excluded.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ src/Libraries/Nop.Core/Domain/Attributes/BaseAttributeValue.cs
@@ -25,5 +25,5 @@ public abstract partial class BaseAttributeValue : BaseEntity, ILocalizedEntity
     /// <summary>
     /// Gets or sets the attribute identifier
     /// </summary>
-    public int AttributeId { get; set; }
+    public long AttributeId { get; set; }
 }
