+++ src/Libraries/Nop.Core/Domain/Catalog/CrossSellProduct.cs
@@ -8,10 +8,10 @@ public partial class CrossSellProduct : BaseEntity
     /// <summary>
     /// Gets or sets the first product identifier
     /// </summary>
-    public int ProductId1 { get; set; }
+    public long ProductId1 { get; set; }
 
     /// <summary>
     /// Gets or sets the second product identifier
     /// </summary>
-    public int ProductId2 { get; set; }
+    public long ProductId2 { get; set; }
 }
