+++ src/Libraries/Nop.Core/Domain/Catalog/ProductPicture.cs
@@ -8,12 +8,12 @@ public partial class ProductPicture : BaseEntity
     /// <summary>
     /// Gets or sets the product identifier
     /// </summary>
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
 
     /// <summary>
     /// Gets or sets the picture identifier
     /// </summary>
-    public int PictureId { get; set; }
+    public long PictureId { get; set; }
 
     /// <summary>
     /// Gets or sets the display order
