+++ src/Libraries/Nop.Core/Domain/Catalog/RelatedProduct.cs
@@ -8,12 +8,12 @@ public partial class RelatedProduct : BaseEntity
     /// <summary>
     /// Gets or sets the first product identifier
     /// </summary>
-    public int ProductId1 { get; set; }
+    public long ProductId1 { get; set; }
 
     /// <summary>
     /// Gets or sets the second product identifier
     /// </summary>
-    public int ProductId2 { get; set; }
+    public long ProductId2 { get; set; }
 
     /// <summary>
     /// Gets or sets the display order
