+++ src/Libraries/Nop.Core/Domain/Catalog/StockQuantityHistory.cs
@@ -28,15 +28,15 @@ public partial class StockQuantityHistory : BaseEntity
     /// <summary>
     /// Gets or sets the product identifier
     /// </summary>
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
 
     /// <summary>
     /// Gets or sets the product attribute combination identifier
     /// </summary>
-    public int? CombinationId { get; set; }
+    public long? CombinationId { get; set; }
 
     /// <summary>
     /// Gets or sets the warehouse identifier
     /// </summary>
-    public int? WarehouseId { get; set; }
+    public long? WarehouseId { get; set; }
 }
