+++ src/Libraries/Nop.Core/Domain/Common/AddressSettings.cs
@@ -75,7 +75,7 @@ public partial class AddressSettings : ISettings
     /// <summary>
     /// Gets or sets a Default Country
     /// </summary>
-    public int? DefaultCountryId { get; set; }
+    public long? DefaultCountryId { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether 'State / province' is enabled
