+++ src/Libraries/Nop.Core/Domain/Common/GenericAttribute.cs
@@ -8,7 +8,7 @@ public partial class GenericAttribute : BaseEntity
     /// <summary>
     /// Gets or sets the entity identifier
     /// </summary>
-    public int EntityId { get; set; }
+    public long EntityId { get; set; }
 
     /// <summary>
     /// Gets or sets the key group
@@ -28,7 +28,7 @@ public partial class GenericAttribute : BaseEntity
     /// <summary>
     /// Gets or sets the store identifier
     /// </summary>
-    public int StoreId { get; set; }
+    public long StoreId { get; set; }
 
     /// <summary>
     /// Gets or sets the created or updated date
