+++ src/Libraries/Nop.Core/Domain/Common/SearchTerm.cs
@@ -13,12 +13,12 @@ public partial class SearchTerm : BaseEntity, ISoftDeletedEntity
     /// <summary>
     /// Gets or sets the store identifier
     /// </summary>
-    public int StoreId { get; set; }
+    public long StoreId { get; set; }
 
     /// <summary>
     /// Gets or sets the customer identifier
     /// </summary>
-    public int CustomerId { get; set; }
+    public long CustomerId { get; set; }
 
     /// <summary>
     /// Gets or sets the date and time of instance creation
