+++ src/Libraries/Nop.Core/Domain/Customers/BestCustomerReportLine.cs
@@ -8,7 +8,7 @@ public partial class BestCustomerReportLine
     /// <summary>
     /// Gets or sets the customer identifier
     /// </summary>
-    public int CustomerId { get; set; }
+    public long CustomerId { get; set; }
     /// <summary>
     /// Gets or sets the customer name
     /// </summary>
