+++ src/Libraries/Nop.Core/Domain/Customers/CustomerPassword.cs
@@ -13,7 +13,7 @@ public partial class CustomerPassword : BaseEntity
     /// <summary>
     /// Gets or sets the customer identifier
     /// </summary>
-    public int CustomerId { get; set; }
+    public long CustomerId { get; set; }
 
     /// <summary>
     /// Gets or sets the password
@@ -23,7 +23,7 @@ public partial class CustomerPassword : BaseEntity
     /// <summary>
     /// Gets or sets the password format identifier
     /// </summary>
-    public int PasswordFormatId { get; set; }
+    public long PasswordFormatId { get; set; }
 
     /// <summary>
     /// Gets or sets the password salt
