+++ src/Libraries/Nop.Core/Domain/Customers/RewardPointsHistory.cs
@@ -8,12 +8,12 @@ public partial class RewardPointsHistory : BaseEntity
     /// <summary>
     /// Gets or sets the customer identifier
     /// </summary>
-    public int CustomerId { get; set; }
+    public long CustomerId { get; set; }
 
     /// <summary>
     /// Gets or sets the store identifier in which these reward points were awarded or redeemed
     /// </summary>
-    public int StoreId { get; set; }
+    public long StoreId { get; set; }
 
     /// <summary>
     /// Gets or sets the points redeemed/added
