+++ src/Libraries/Nop.Core/Domain/Discounts/DiscountCategoryMapping.cs
@@ -8,5 +8,5 @@ public partial class DiscountCategoryMapping : DiscountMapping
     /// <summary>
     /// Gets or sets the category identifier
     /// </summary>
-    public override int EntityId { get; set; }
+    public override long EntityId { get; set; }
 }
