+++ src/Libraries/Nop.Core/Domain/Discounts/DiscountMapping.cs
@@ -5,15 +5,15 @@ public abstract partial class DiscountMapping : BaseEntity
     /// <summary>
     /// Gets the entity identifier
     /// </summary>
-    public new int Id { get; }
+    public new long Id { get; }
 
     /// <summary>
     /// Gets or sets the discount identifier
     /// </summary>
-    public int DiscountId { get; set; }
+    public long DiscountId { get; set; }
 
     /// <summary>
     /// Gets or sets the entity identifier
     /// </summary>
-    public abstract int EntityId { get; set; }
+    public abstract long EntityId { get; set; }
 }
