+++ src/Libraries/Nop.Core/Domain/Discounts/DiscountRequirement.cs
@@ -8,7 +8,7 @@ public partial class DiscountRequirement : BaseEntity
     /// <summary>
     /// Gets or sets the discount identifier
     /// </summary>
-    public int DiscountId { get; set; }
+    public long DiscountId { get; set; }
 
     /// <summary>
     /// Gets or sets the discount requirement rule system name
@@ -18,12 +18,12 @@ public partial class DiscountRequirement : BaseEntity
     /// <summary>
     /// Gets or sets the parent requirement identifier
     /// </summary>
-    public int? ParentId { get; set; }
+    public long? ParentId { get; set; }
 
     /// <summary>
     /// Gets or sets an interaction type identifier (has a value for the group and null for the child requirements)
     /// </summary>
-    public int? InteractionTypeId { get; set; }
+    public long? InteractionTypeId { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether this requirement has any child requirements
