+++ src/Libraries/Nop.Core/Domain/FilterLevels/FilterLevelValueProductMapping.cs
@@ -8,11 +8,11 @@ public partial class FilterLevelValueProductMapping : BaseEntity
     /// <summary>
     /// Gets or sets the filter level value identifier
     /// </summary>
-    public int FilterLevelValueId { get; set; }
+    public long FilterLevelValueId { get; set; }
     
     /// <summary>
     /// Gets or sets the product identifier
     /// </summary>
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
     
 }
