+++ src/Libraries/Nop.Core/Domain/Localization/LocaleStringResource.cs
@@ -8,7 +8,7 @@ public partial class LocaleStringResource : BaseEntity
     /// <summary>
     /// Gets or sets the language identifier
     /// </summary>
-    public int LanguageId { get; set; }
+    public long LanguageId { get; set; }
 
     /// <summary>
     /// Gets or sets the resource name
