+++ src/Libraries/Nop.Core/Domain/Messages/EmailAccountSettings.cs
@@ -10,5 +10,5 @@ public partial class EmailAccountSettings : ISettings
     /// <summary>
     /// Gets or sets a store default email account identifier
     /// </summary>
-    public int DefaultEmailAccountId { get; set; }
+    public long DefaultEmailAccountId { get; set; }
 }
